[backend] refactor #35   - モデル定義をファイル分け
