+++ backend/internal/model/user.go
@@ -22,29 +22,6 @@ type User struct {
 	CreatedAt time.Time `gorm:"not null"`
 	UpdatedAt time.Time `gorm:"not null"`
 
-	Profile   Profile   `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
-	Works     []Work    `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
-}
-
-type Profile struct {
-	ID        uint      `gorm:"primaryKey;autoIncrement"`
-	UserID    uuid.UUID `gorm:"type:uuid;not null;uniqueIndex"`
-	LastName  string    `gorm:"not null"`
-	FirstName string    `gorm:"not null"`
-	BirthDate time.Time
-	School    string
-	Hobby     string
-	Skills    []Skill   `gorm:"foreignKey:ProfileID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
-	CreatedAt time.Time `gorm:"not null"`
-	UpdatedAt time.Time `gorm:"not null"`
-}
-
-type Skill struct {
-	ID          uint      `gorm:"primaryKey;autoIncrement"`
-	ProfileID   uint      `gorm:"index;not null"`
-	Name        string    `gorm:"not null"`
-	Description string
-	IsMainSkill bool
-	CreatedAt   time.Time `gorm:"not null"`
-	UpdatedAt   time.Time `gorm:"not null"`
+	Profile Profile `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
+	Works   []Work  `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
 }
